初期化処理とウィンドウ設定を分離
